macOS: Hard-code NEON and SSE/SSE2 when building universal builds
We don't have the configure machinery to run two sets of arch tests,
which means all the CPU features are based on the primary architecture.
Like for iOS, we hard-code the feature of the second architecture, but
limited to NEON and SSE.

If the primary architecture is arm64, we'll miss a bunch of x86_64
features (locally I have cx16 mmx sse sse2 sse3 ssse3 sse4.1). We
are also leaving out the arm64 crc32 feature.

Change-Id: Idf384e413a3ae7ab2e533217db36f58dafb144e4
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>