Never handle scrollbars styled with box or border changes as transient
If scrollbars are styled with a style sheet that includes a box or
removes off the native border, then we never treat them as transient or
overlapping. Otherwise, the layout logic in QAbstractScrollArea will
show them on top of the viewport, overlapping the content.

The cherry-pick doesn't include the base line test from the upstream
change, as the testing framework is not available in 6.2.

Fixes: QTBUG-98289
Change-Id: Ic53ad248b3eedd54722ac7b2fe5256a27092dbc1
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 56da533b45e2785e8fd54ef6fcddb7b973c665bf)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>